Split the logic for pylons and flask plugin
