+++ ckanext/datarequests/plugins/flask_plugin.py
@@ -1,0 +1,11 @@
+import ckan.plugins as p
+import ckanext.datarequests.views as views
+
+
+class MixinPlugin(p.SingletonPlugin):
+    p.implements(p.IBlueprint)
+
+    # IBlueprint
+
+    def get_blueprint(self):
+        return views.get_blueprints()
