+++ ckanext/datarequests/validator.py
@@ -17,10 +17,10 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with CKAN Data Requests Extension. If not, see <http://www.gnu.org/licenses/>.
 
-import constants
-import ckan.plugins.toolkit as tk
-import ckanext.datarequests.db as db
 
+import ckan.plugins.toolkit as tk
+import ckanext.datarequests.constants as constants
+import ckanext.datarequests.db as db
 
 def validate_datarequest(context, request_data):
